Reorganize source structure
